Make Api to use encrypted Data
