+++ packages/shared/config/environment/index.ts
@@ -1,6 +1,7 @@
-export default {
+export const ServerConfig = {
   encKey: '',
   encrypted: {
-    mongoUrl: 'KgvapgxntZoNpyh9aVgeb0Ny2y02XbeEPXOw4xKnXxPvROVP'
+    mongoUrl:
+      'bBwbf3kPmjpObNvbaJaDG60MR7HnLEA6Wfl1KnnXf1VdaI9bIjLkMAPtXQmbQZVZ82z5E4oWa3s9SNIUqQXhgwcy9nHY2wBwSmbf00NAid+vQguAlH1jLgEU85WdrGpr7ewnfTesDOJP9yKK+FCCKrUwdQ=='
   }
 }
