+++ packages/shared/config/index.ts
@@ -1,0 +1,2 @@
+export * from './paths'
+export * from './environment'
